Added a cast to an int when fscanf()'ing into errType, which is really an enum, to make insure++ happy.

+++ src/condor_c++_util/condor_event.C
@@ -353,7 +353,7 @@ readEvent (FILE *file)
 	char buffer [128];
 
 	// get the error number
-	retval = fscanf (file, "(%d)", &errType);
+	retval = fscanf (file, "(%d)", &(int)errType);
 	if (retval != 1) 
 	{ 
 		return 0;
